// angular-loan-frontend/src/app/models/LoanDetails.model.ts
export class LoanDetails{
    userId: string;
    annualSalary: number;
    loanAmount: number;
    duration: number;
    loanType: string;
    differentAssest: string;
    assestValuation: number;
    referPerson: string;
    finalAmount: number;
    numberofEmi: number;
    sanctionDate: string;
    paidEmi: number;
    lateCharges: number;
    loanApproval: boolean;
}